✨ [FEAT] 홈 화면 즐겨찾기 등록 api 구현 #74
