+++ src/main/java/org/sopt/lequuServer/domain/book/facade/BookFacade.java
@@ -2,13 +2,10 @@ package org.sopt.lequuServer.domain.book.facade;
 
 import lombok.RequiredArgsConstructor;
 import org.sopt.lequuServer.domain.book.dto.request.BookCreateRequestDto;
-import org.sopt.lequuServer.domain.book.dto.request.FavoriteCreateRequestDto;
 import org.sopt.lequuServer.domain.book.dto.response.BookCreateResponseDto;
 import org.sopt.lequuServer.domain.book.dto.response.BookDetailResponseDto;
 import org.sopt.lequuServer.domain.book.model.Book;
-import org.sopt.lequuServer.domain.book.model.Favorite;
 import org.sopt.lequuServer.domain.book.repository.BookRepository;
-import org.sopt.lequuServer.domain.book.repository.FavoriteRepository;
 import org.sopt.lequuServer.domain.book.service.BookService;
 import org.sopt.lequuServer.domain.member.model.Member;
 import org.sopt.lequuServer.domain.member.repository.MemberRepository;
@@ -40,7 +37,6 @@ public class BookFacade {
     private final PostedStickerRepository postedStickerRepository;
     private final BadWordFilterService badWordFilterService;
     private final S3Service s3Service;
-    private final FavoriteRepository favoriteRepository;
 
     @Transactional
     public BookCreateResponseDto createBook(BookCreateRequestDto request, Long memberId) {
@@ -106,15 +102,4 @@ public class BookFacade {
         return BookDetailResponseDto.of(book);
     }
 
-    @Transactional
-    public void createFavorite(Long memberId, FavoriteCreateRequestDto request) {
-        Member member = memberRepository.findByIdOrThrow(memberId);
-        Book book = bookRepository.findByIdOrThrow(request.bookId());
-
-        Favorite favorite = Favorite.of(member, book);
-        favoriteRepository.save(favorite);
-
-        book.addFavorite(favorite);
-        member.addFavorite(favorite);
-    } // memberId와 bookId를 favorite에 저장하는 로직
 }
