+++ src/main/java/org/sopt/lequuServer/domain/book/model/Book.java
@@ -2,6 +2,7 @@ package org.sopt.lequuServer.domain.book.model;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.sopt.lequuServer.domain.favorite.model.Favorite;
 import org.sopt.lequuServer.domain.note.model.Note;
 import org.sopt.lequuServer.domain.sticker.model.PostedSticker;
 import org.sopt.lequuServer.domain.member.model.Member;
