+++ src/main/java/org/sopt/lequuServer/domain/favorite/dto/request/FavoriteCreateRequestDto.java
@@ -1,4 +1,4 @@
-package org.sopt.lequuServer.domain.book.dto.request;
+package org.sopt.lequuServer.domain.favorite.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
