+++ src/main/java/org/sopt/lequuServer/domain/favorite/repository/FavoriteRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.lequuServer.domain.book.repository;
+package org.sopt.lequuServer.domain.favorite.repository;
 
-import org.sopt.lequuServer.domain.book.model.Favorite;
+import org.sopt.lequuServer.domain.favorite.model.Favorite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
